Dry up different classes and inherit from two parent classes
